+++ templates/admin-panel/template/client/src/storages/session.js
@@ -1,6 +1,6 @@
 import createStorage from 'sham-ui-data-storage';
 
-export const { useStorage } = createStorage( {
+export const { storage, useStorage } = createStorage( {
     name: '',
     email: '',
     sessionValidated: false,